Update AssemblyScript, WASA and the AssemblyScript example and instructions
This is an update to the AssemblyScript module and example, after the new
runtimes landed into the AssemblyScript compiler.

WASA, the WASI bindings, got a major update, not only to support the new ASC,
but also to support most of WASI as it is today.
Breaking changes were made to the API to make is both safer and closer to the
NodeJS filesystem API.

wasm-crypto was updated for the new runtime as well.

The example was updated for the new runtime and the new WASA API.

The AssemblyScript compiler still has some rough edges regarding garbage
collection. So, switch the runtime from `full` (default) to `stub` for now,
until some known regressions have been ironed out.